lib/rbloom/benchmarks/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile ./requirements.in
-e ../
    # via -r ./requirements.in
bitarray==3.7.1
    # via pybloom3
bloom-filter2==2.0.0
    # via -r ./requirements.in
flor==1.1.3
    # via -r ./requirements.in
pybloom3==0.0.3
    # via -r ./requirements.in
pybloomfiltermmap3==0.6.0
    # via -r ./requirements.in
